+++ lib/features/steps/app_automator_steps.rb
@@ -53,26 +53,75 @@ When("I send the keys {string} to the element {string}") do |keys, element_id|
 end
 
 # Tests that the given payload value is correct for the target BrowserStack platform.
+# This step will assume the expected and payload values are strings.
 # If the step is invoked when a remote BrowserStack device is not in use this step will fail.
 #
 # The DataTable used for this step should have `ios` and `android` in the same row as their expected value:
 #   | android | Java.lang.RuntimeException |
 #   | ios     | NSException                |
 #
-# If the expected value is set to "skip", the check should be skipped.
+# If the expected value is set to "@skip", the check should be skipped.
 #
 # @step_input field_path [String] The field to test
 # @step_input platform_values [DataTable] A table of acceptable values for each platform
-Then("the event {string} matches the correct platform value:") do |field_path, platform_values|
-  if !defined?($driver) || $driver.nil?
-    fail("This step should only be used if the AppAutomateDriver is present")
-  end
-  os = $driver.capabilities['os']
-  expected_value = Hash[platform_values.raw][os]
-  fail("There is no expected value for the current platform \"#{os}\"") if expected_value.nil?
-  unless expected_value.eql?("@skip")
-    assert_equal(expected_value, read_key_path(Server.current_request[:body], "events.0.#{field_path}"))
-  end
+Then("the payload field {string} equals the platform-dependent string:") do |field_path, platform_values|
+  testStringPlatformValues(field_path, platform_values)
+end
+
+# See `the payload field {string} equals the platform-dependent string:`
+#
+# @step_input field_path [String] The field to test, prepended with "events.0"
+# @step_input platform_values [DataTable] A table of acceptable values for each platform
+Then("the event {string} equals the platform-dependent string:") do |field_path, platform_values|
+  testStringPlatformValues("events.0.#{field_path}", platform_values)
+end
+
+# Tests that the given payload value is correct for the target BrowserStack platform.
+# This step will assume the expected and payload values are numeric.
+# If the step is invoked when a remote BrowserStack device is not in use this step will fail.
+#
+# The DataTable used for this step should have `ios` and `android` in the same row as their expected value:
+#   | android | 1  |
+#   | ios     | 5.5 |
+#
+# If the expected value is set to "@skip", the check should be skipped.
+#
+# @step_input field_path [String] The field to test
+# @step_input platform_values [DataTable] A table of acceptable values for each platform
+Then("the payload field {string} equals the platform-dependent numeric:") do |field_path, platform_values|
+  testNumericPlatformValues(field_path, platform_values)
+end
+
+# See `the payload field {string} equals the platform-dependent numeric:`
+#
+# @step_input field_path [String] The field to test, prepended with "events.0"
+# @step_input platform_values [DataTable] A table of acceptable values for each platform
+Then("the event {string} equals the platform-dependent numeric:") do |field_path, platform_values|
+  testNumericPlatformValues("events.0.#{field_path}", platform_values)
+end
+
+# Tests that the given payload value is correct for the target BrowserStack platform.
+# This step will assume the expected and payload values are booleans.
+# If the step is invoked when a remote BrowserStack device is not in use this step will fail.
+#
+# The DataTable used for this step should have `ios` and `android` in the same row as their expected value:
+#   | android | 1 |
+#   | ios     | 5 |
+#
+# If the expected value is set to "@skip", the check should be skipped.
+#
+# @step_input field_path [String] The field to test
+# @step_input platform_values [DataTable] A table of acceptable values for each platform
+Then("the payload field {string} equals the platform-dependent boolean:") do |field_path, platform_values|
+  testBooleanPlatformValues(field_path, platform_values)
+end
+
+# See `the payload field {string} equals the platform-dependent boolean:`
+#
+# @step_input field_path [String] The field to test, prepended with "events.0"
+# @step_input platform_values [DataTable] A table of acceptable values for each platform
+Then("the event {string} equals the platform-dependent boolean:") do |field_path, platform_values|
+  testBooleanPlatformValues("events.0.#{field_path}", platform_values)
 end
 
 # Sends keys to a given element, clearing it first
@@ -82,4 +131,49 @@ end
 # @step_input element_id [String] The locator id
 When("I clear and send the keys {string} to the element {string}") do |keys, element_id|
   $driver.clear_and_send_keys_to_element(element_id, keys)
+end
+
+def getExpectedValueForPlatform(platform_values)
+  if !defined?($driver) || $driver.nil?
+    fail("This step should only be used if the AppAutomateDriver is present")
+  end
+  os = $driver.capabilities['os']
+  expected_value = Hash[platform_values.raw][os]
+  fail("There is no expected value for the current platform \"#{os}\"") if expected_value.nil?
+  expected_value
+end
+
+def shouldSkipPlatformCheck(expected_value)
+  expected_value.eql?("@skip")
+end
+
+def testStringPlatformValues(field_path, platform_values)
+  expected_value = getExpectedValueForPlatform(platform_values)
+  unless shouldSkipPlatformCheck(expected_value)
+    payload_value = read_key_path(Server.current_request[:body], field_path)
+    assert_equal(expected_value, payload_value)
+  end
+end
+
+def testBooleanPlatformValues(field_path, platform_values)
+  expected_value = getExpectedValueForPlatform(platform_values)
+  unless shouldSkipPlatformCheck(expected_value)
+    if expected_value.downcase == 'true'
+      expected_bool = true
+    elsif expected_value.downcase == 'false'
+      expected_bool = false
+    else
+      expected_bool = expected_value
+    end
+    payload_value = read_key_path(Server.current_request[:body], field_path)
+    assert_equal(expected_bool, payload_value)
+  end
+end
+
+def testNumericPlatformValues(field_path, platform_values)
+  expected_value = getExpectedValueForPlatform(platform_values)
+  unless shouldSkipPlatformCheck(expected_value)
+    payload_value = read_key_path(Server.current_request[:body], field_path)
+    assert_equal(expected_value.to_f, payload_value)
+  end
 end