+++ test/fixtures/browserstack-app/app/src/main/java/com/bugsnag/android/mazerunner/MainActivity.kt
@@ -5,8 +5,7 @@ import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import android.util.Log
 import android.widget.Button
-import com.bugsnag.android.Bugsnag
-import com.bugsnag.android.Configuration
+import com.bugsnag.android.*
 import java.lang.Exception
 
 class MainActivity : AppCompatActivity() {
@@ -15,7 +14,15 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
         val button = findViewById<Button>(R.id.trigger_error)
-        button.setOnClickListener { Bugsnag.notify(Exception("HandledException!")) }
+        button.setOnClickListener { 
+            Bugsnag.notify(Exception("HandledException!"), {
+                val error = it.error!!
+                error.metaData.addToTab("test", "boolean_false", false)
+                error.metaData.addToTab("test", "boolean_true", true)
+                error.metaData.addToTab("test", "float", 1.55)
+                error.metaData.addToTab("test", "integer", 2)
+            })
+        }
     }
 
     override fun onResume() {